+++ utils/stream-utils.h
@@ -5,7 +5,7 @@
 
 long fsize(FILE *file);
 long fleft(FILE *file);
-long fleftat(int pos, FILE *file);
-int fmove(int pos, int offset, FILE *file);
+long fleftat(unsigned long int pos, FILE *file);
+int fmove(unsigned long int pos, long int offset, FILE *file);
 
 #endif
